bcm63xx: Add board id override fixup.
Add a fixup for the kernel to check the bcm tag of the image for a boardid in
the information1 field, and use that instead of the supplied one by cfe. This
is triggered by a + at the start of the information1 field.

Also modify the image generation to generate apropriately tagged images.

Signed-off-by: Jonas Gorski <jonas.gorski+openwrt@gmail.com>